backStack

+++ app/src/main/java/kz/sherua/nadoprodat/model/dbentity/SaleDetails.kt
@@ -11,7 +11,7 @@ data class SaleDetails(
     val salesId: Long,
     val productId: Long,
     val productSalesPrice: Double,
-    val productCount: Long,
+    val productCount: Int,
     val currencyId: Long? = null,
     val crDate: Long,
     val upDate: Long? = null
